+++ src/main/java/br/com/personalog/dao/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface UserDao extends JpaRepository<User, Integer>, UserDaoQuery {
+public interface UserDao extends JpaRepository<User, Long>, UserDaoQuery {
 
 	@Query(value = "SELECT CASE WHEN COUNT(u) > 0 THEN true ELSE false END FROM User u WHERE u.email = :email")
 	Boolean isEmailExists(@Param("email") String email);
